Add crypto
This commit makes the communication between server and client encrypted.
Be careful, as there are other potential leaks of information:
- message timing
- message size (variable length encoded integers!)
- communication with domjudge (http(s)!)

The current design is:
Server and Client share a secret password. A key is derived with scrypt
and a random salt at server startup. The server sends the salt together
with a (per client) 16 byte nonce to the client. NaCL secretbox
(https://nacl.cr.yp.to/secretbox.html) is then used to encrypt (and
authenticate) all further communication (apart from packet sizes) with
the shared key (the nonce is an 8 byte counter ++ server generated
initial nonce).